New interpolation functions to apply boundary conditions
This commit has a proof-of-concept prototype where we have implemented three different interpolation functions that take into account the position of the particle with respect to land. No slip, Free slip and Free slip with no-normal flow.

The peninsula_example has been adapted to show how the new methods work.

The eval method of the class Field is the one that uses the interpolation. In order to try ours the 'self.spatial_interpolation' calls have to be changed.

Moreover, the class Particle now has an extra attribute, dt_initial that is intended to hold the original dt (at the time the particle is created). We use this to change the dt when the particle is close to land and restore it to its original value when is not.